+++ src/firebase-messaging-sw.js
@@ -2,47 +2,12 @@
 // These scripts are made available when the app is served or deployed on Firebase Hosting
 // If you do not serve/host your project using Firebase Hosting see https://firebase.google.com/docs/web/setup
 
-// ------ old ------
-// importScripts('https://www.gstatic.com/firebasejs/5.8.6/firebase-app.js');
-// importScripts('https://www.gstatic.com/firebasejs/5.8.6/firebase-messaging.js');
-// importScripts('https://www.gstatic.com/firebasejs/5.8.6/firebase.js');
-
-// ------ new ------
-// import * as firebase from 'firebase/app';
-// import '@firebase/messaging';
 
 importScripts('https://www.gstatic.com/firebasejs/8.6.7/firebase-app.js');
 importScripts('https://www.gstatic.com/firebasejs/8.6.7/firebase-messaging.js');
-// importScripts('https://www.gstatic.com/firebasejs/8.6.7/firebase.js');
 
 
-
-// importScripts('/__/firebase/3.9.0/firebase-app.js');
-// importScripts('/__/firebase/3.9.0/firebase-messaging.js');
-// importScripts('/__/firebase/init.js');
-
-//const messaging = firebase.messaging();
-
-/**
- * Here is is the code snippet to initialize Firebase Messaging in the Service
- * Worker when your app is not hosted on Firebase Hosting.
- **/
- // [START initialize_firebase_in_sw]
- // Give the service worker access to Firebase Messaging.
- // Note that you can only use Firebase Messaging here, other Firebase libraries
- // are not available in the service worker.
- //importScripts('https://www.gstatic.com/firebasejs/3.9.0/firebase-app.js');
- //importScripts('https://www.gstatic.com/firebasejs/3.9.0/firebase-messaging.js');
- // Initialize the Firebase app in the service worker by passing in the
- // messagingSenderId.
-//  firebase.initializeApp({
-//    'messagingSenderId': '77360455507' // old
-// //  'messagingSenderId': '269505353043'
-  
-//  });
-
-
- firebase.initializeApp({
+firebase.initializeApp({
   apiKey: 'AIzaSyCoWXHNvP1-qOllCpTshhC6VjPXeRTK0T4"',
   authDomain: 'chat21-pre-01.firebaseapp.com',
   databaseURL: 'https://chat21-pre-01.firebaseio.com',
@@ -51,56 +16,85 @@ importScripts('https://www.gstatic.com/firebasejs/8.6.7/firebase-messaging.js');
   messagingSenderId: '269505353043',
   appId: '1:269505353043:web:b82af070572669e3707da6'
 });
- // Retrieve an instance of Firebase Messaging so that it can handle background
- // messages.
- const messaging = firebase.messaging();//.useServiceWorker(registration);
- 
- // [END initialize_firebase_in_sw]
- 
+// Retrieve an instance of Firebase Messaging so that it can handle background
+// messages.
+const messaging = firebase.messaging();//.useServiceWorker(registration);
 
+// [END initialize_firebase_in_sw]
 
-// If you would like to customize notifications that are received in the
-// background (Web app is closed or not in browser focus) then you should
-// implement this optional method.
-// [START background_handler]
-
-// ----- old -------
-// messaging.setBackgroundMessageHandler(function(payload) {
-//   console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) Received background message ', payload);
-//   // Customize notification here
-//   const notificationTitle = payload.notification.title;
-//   const notificationOptions = {
-//     body: payload.notification.body,
-//     // icon: '/firebase-logo.png'
-//   };
-
-//   return self.registration.showNotification(notificationTitle,
-//       notificationOptions);
-// });
-
-// Handle incoming messages. Called when:
-// - a message is received while the app has focus
-// - the user clicks on an app notification created by a service worker
-//   `messaging.onBackgroundMessage` handler.
-// messaging.onMessage((payload) => {
-//   console.log('FIREBASE-NOTIFICATION FIREBASE-MESSAGING-SW) Message received. ', payload);
-//   // ...
-// });
 
 
 // ----- new -------
-messaging.onBackgroundMessage(function(payload) {
-  console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) Received background message ', payload);
-  
+messaging.onBackgroundMessage(function (payload) {
+  // console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) Received background message ', payload);
+  // console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) Received background message payload recipient', payload.data.recipient);
+
+  data = `{ "recipient": "${payload.data.recipient}", "recipient_fullname": "${payload.data.recipient_fullname}", "status": "${payload.data.status}" }`
   const notificationTitle = payload.notification.title;
   const notificationOptions = {
-  body: payload.notification.body,
-  icon: './assets/images/tiledesk_logo_50x50.png',
-  data: { url:payload.data.click_action }, //the url which we gonna use later
-  actions: [{action: "open_url", title: "Read Now"}]
+    body: payload.notification.body,
+    icon: './assets/images/tiledesk_logo_50x50.png',
+    data: { url: payload.data.click_action }, //the url which we gonna use later
+    actions: [{
+      action: data, title: "Read Now"
+    }]
   };
   // /Users/nicola/CHAT21_IONIC/src/assets/images/tiledesk_logo_no_text_72x72.png
-  self.registration.showNotification(notificationTitle,
-  notificationOptions);
-  });
+  self.registration.showNotification(notificationTitle, notificationOptions);
+});
 // [END background_handler]
+
+
+self.addEventListener('notificationclick', function (event) {
+  // console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) notificationclick  +++ event ', event);
+
+  let url = event.notification.data.url;
+  // let url = "http://localhost:8101/#/conversation-detail/support-group-60aa0ffd1482fe00346854be-8ff54a0e11674e269fab981603b16745/WALTER/active"
+  event.preventDefault();
+  // console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) notificationclick  ++++ event url', url);
+  // console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) notificationclick  ++++ event notification actions > action', event.notification.actions[0]['action']);
+  
+  // Android doesn’t close the notification when you click on it
+  // See: http://crbug.com/463146
+  event.notification.close();
+
+// This looks to see if the current is already open and
+  // focuses if it is
+event.waitUntil(self.clients.matchAll(
+  {
+    includeUncontrolled: true,
+    type: 'window'
+  }).then(function (clientList) {
+    for (var i = 0; i < clientList.length; i++) {
+      var client = clientList[i];
+      // console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) notificationclick client ', client);
+      if (client.url.startsWith(url) && 'focus' in client) {
+      
+        // -----------------------------------------------
+        // Post message
+        // -----------------------------------------------
+        client.postMessage({
+          message: 'Received a push message.',
+          data: event.notification.actions[0]['action']
+        });
+
+        // console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) HERE YES 1 ');
+        return client.focus();
+      }
+    }
+    if (clients.openWindow) {
+      console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) HERE YES  2');
+      console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) HERE YES  2  notification  actions ', event.notification.actions[0]['action']);
+       const  actioObjct = JSON.parse(event.notification.actions[0]['action']);
+      console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) HERE YES  2  notification  actioObjct ', actioObjct);
+      // '#/conversation-detail/' +
+      const url = event.notification.data.url +  actioObjct.recipient + "/" + actioObjct.recipient_fullname + '/active'
+      console.log('FIREBASE-NOTIFICATION (FIREBASE-MESSAGING-SW) HERE YES  2  built url ', url);
+
+   
+      return clients.openWindow(url);
+     
+    }
+  }));
+});
+